pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iac-checker"
version = "0.2.0"
description = "Azure IaC compliance checker (Terraform, ARM, Bicep) against WAF & CAF best practices"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "python-hcl2>=4.3.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
iac-checker = "iac_checker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
